fixed calculations
