+++ src/Pages/SavingSuggestions.cshtml
@@ -80,7 +80,7 @@
         </p>
         @if (!Model.WasException)
         {
-            <h4 class="text-light mt-2 mb-2">Income: @NumberFormatter.FormatCurrency(Model.MonthlyIncome, Model.CurrencySymbol)</h4>
+            <h4 class="text-light mt-2 mb-2">Monthly Income: @NumberFormatter.FormatCurrency(Model.MonthlyIncome, Model.CurrencySymbol)</h4>
 
         }
         <hr />
